Move search in map as a separate function
